Do not include spdl.h from RcppSpdlog
For now do not use vararg logger via spdl.h which has been seen to cause side
effects at least under static linking in Windows. The plan is to (in due
course) bring either a reworked logger interface at the C++ level, or make
do with a more limited one (as the one we started with relying on tinyfmt).

Eventually with C++20 and its included fmt library all will be easier anyway.